+++ src/loaders/settings/scan.tsx
@@ -1,6 +1,6 @@
 import { SCAN_SETTINGS_GROUPED } from "@/lib/constants/settings/scan-options";
 import { ScanOptionGroup } from "@/lib/types/enums";
-import ExclusionsLoader from "@/components/loaders/exclusions";
+import FolderPathFormLoader from "@/components/loaders/path-form";
 import SettingsOptionLoader from "@/components/loaders/settings-options";
 import SettingsItemLoader from "@/components/loaders/settings-item";
 
@@ -17,7 +17,7 @@ export default function ScanSettingsLoader(){
                          ))}
                     </SettingsItemLoader>
                ))}
-               <ExclusionsLoader/>
+               <FolderPathFormLoader/>
           </div>
      )
 }